Arc_Prize/requirements.txt
streamlit==1.29.0
numpy==1.26.0
matplotlib==3.8.0
pandas==2.1.1
pillow==10.0.1